(#8) Apply all rebar.config files
So far we found the closest rebar.config file when walking up the directory
tree from the source file and used that config file.

Now we always walk up until we reach the root directory and use all
rebar.config files to find the include files and BEAM files. We still use only
the closes rebar.config file to determine the other compilation options,
because that's what rebar seems to do.