Flymake: don't check the files twice
When flymake called erlang_check.erl, it passed the name of the module
twice:

    :!.../vim-erlang-compiler/compiler/erlang_check.erl --outdir .
     mymod.erl 'mymod.erl'

This caused erlang_check to check this file twice, which doubled its
execution time.

The reason was that the AutoRun function added the current file to
makeprg, and then it also added it when calling "silent lmake!".

The first of these additions is removed by this commit. So now the
check is done only once.